🚑 fix: update special_keyconfig and karabiner.json configurations
This commit modifies the `special_keyconfig.json` file to mark the title as "(⚠️ Disabled) Special keyconfig" and removes a specific rule related to capturing screenshots and copying to the clipboard from the `karabiner.json` file. Additionally, it adds new key mappings for translation and text recognition in the `raycast/README.md` file.